+++ src/extension.ts
@@ -14,6 +14,7 @@ import {
     RuleProvider,
     TypescriptProvider,
 } from "./providers";
+import { CFamilyProvider } from "./providers/cfamily";
 
 function goToDefinition(range: vscode.Range) {
     const editor: vscode.TextEditor = vscode.window.activeTextEditor;
@@ -62,8 +63,12 @@ export function activate(context: vscode.ExtensionContext) {
         providers.push(new CssProvider());
     }
 
-    if (allowedProviders.length === 0 || allowedProviders.indexOf("css") !== -1) {
+    if (allowedProviders.length === 0 || allowedProviders.indexOf("less") !== -1) {
         providers.push(new LessProvider());
+    }
+
+    if (allowedProviders.length === 0 || allowedProviders.indexOf("cfamily") !== -1) {
+        providers.push(new CFamilyProvider());
     }
 
     const provider = new Provider(providers as Array<IBaseProvider<string | vscode.TreeItem>>);