+++ src/provider.ts
@@ -124,16 +124,28 @@ export class Provider implements vscode.TreeDataProvider<TreeItem> {
             .get("abstractCharacter") as string;
     }
 
+    protected get updateOnError(): boolean {
+        return vscode.workspace.getConfiguration("treeview")
+            .get("updateOnError") as boolean;
+    }
+
     public constructor(langProviders: Array<IBaseProvider<any>>) {
         this.langProviders = langProviders;
         vscode.window.onDidChangeActiveTextEditor((ev?: vscode.TextEditor) => {
-            if (vscode.window.activeTextEditor) {
-                this.refresh(vscode.window.activeTextEditor.document);
+            if ((ev && ev.document) || vscode.window.activeTextEditor) {
+                this.refresh((ev || vscode.window.activeTextEditor).document);
             }
         });
-        vscode.workspace.onDidSaveTextDocument((document) => this.refresh(document));
-        vscode.workspace.onDidOpenTextDocument((document) => this.refresh(document));
-        vscode.workspace.onDidCloseTextDocument((document) => this.refresh(document));
+        vscode.workspace.onDidSaveTextDocument((document) => {
+            if (!this.updateOnError && this.hasErrorsInDiagnostic(document)) {
+                return void 0;
+            }
+
+            this.refresh(document);
+        });
+        vscode.window.onDidChangeVisibleTextEditors((ev?: vscode.TextEditor[]) => {
+            this.refresh(ev[0].document);
+        });
     }
 
     public hasSupport(languageId: string) {
@@ -306,8 +318,9 @@ export class Provider implements vscode.TreeDataProvider<TreeItem> {
                                 });
                             });
                     });
+                }, (err) => {
+                    vscode.window.showWarningMessage(err);
                 });
-
             });
         });
     }
@@ -350,18 +363,14 @@ export class Provider implements vscode.TreeDataProvider<TreeItem> {
             if (tree.functions !== undefined) {
                 items.push(new SectionItem(
                     `Functions`,
-                    (tree.functions.length !== 0) ?
-                        vscode.TreeItemCollapsibleState.Expanded : vscode.TreeItemCollapsibleState.Collapsed,
+                    vscode.TreeItemCollapsibleState.Expanded,
                     "functions-section",
                 ));
             }
 
             if (tree.interfaces !== undefined) {
                 for (const cls of tree.interfaces) {
-                    const collapsed: number =
-                        (tree.traits === undefined || tree.traits.length === 0) &&
-                            (tree.classes === undefined || tree.classes.length === 0) &&
-                            tree.interfaces.indexOf(cls) === 0 ?
+                    const collapsed: number = tree.interfaces.indexOf(cls) === 0 ?
                             vscode.TreeItemCollapsibleState.Expanded :
                             vscode.TreeItemCollapsibleState.Collapsed;
 
@@ -380,8 +389,7 @@ export class Provider implements vscode.TreeDataProvider<TreeItem> {
             if (tree.traits !== undefined) {
                 for (const cls of tree.traits) {
                     const collapsed: number =
-                        (tree.classes === undefined || tree.classes.length === 0) &&
-                            tree.traits.indexOf(cls) === 0 ?
+                        tree.traits.indexOf(cls) === 0 ?
                             vscode.TreeItemCollapsibleState.Expanded :
                             vscode.TreeItemCollapsibleState.Collapsed;
 
@@ -673,5 +681,22 @@ export class Provider implements vscode.TreeDataProvider<TreeItem> {
         }
 
         return items;
+    }
+
+    private hasErrorsInDiagnostic(document: vscode.TextDocument): boolean {
+        const halt = vscode.languages.getDiagnostics().find((x) => {
+            if (x[0].fsPath === document.uri.fsPath) {
+                const diag = x[1].find((y) => {
+                    if (y.severity === vscode.DiagnosticSeverity.Error) {
+                        return true;
+                    }
+                });
+
+                return (diag !== undefined);
+            }
+            return false;
+        });
+
+        return halt !== undefined;
     }
 }